Clean prod code
