+++ core/src/config/test.rs
@@ -1,6 +1,11 @@
+use crate::builder::transaction::output::UnspentTxOut;
+use crate::builder::transaction::TxHandlerBuilder;
+use crate::constants::MIN_TAPROOT_AMOUNT;
 use crate::deposit::DepositData;
 use bitcoin::secp256k1::PublicKey;
 use bitcoin::secp256k1::SecretKey;
+use bitcoin::ScriptBuf;
+use bitcoin::TxOut;
 use bitvm::chunk::api::Assertions;
 use serde::{Deserialize, Serialize};
 use std::str::FromStr;
@@ -245,5 +250,127 @@ impl Default for TestParams {
             generate_varying_total_works_insufficient_total_work: false,
             generate_varying_total_works: false,
         }
+    }
+}
+
+impl TestParams {
+    pub fn maybe_dump_bridge_circuit_params_to_file(
+        &self,
+        bridge_circuit_host_params: &impl borsh::BorshSerialize,
+    ) -> eyre::Result<()> {
+        use std::path::PathBuf;
+
+        if self.use_small_annex {
+            let path = PathBuf::from(
+                "../bridge-circuit-host/bin-files/bch_params_challenge_tx_with_annex.bin",
+            );
+            std::fs::create_dir_all(path.parent().unwrap())
+                .map_err(|e| eyre::eyre!("Failed to create directory for output file: {}", e))?;
+            let serialized_params = borsh::to_vec(&bridge_circuit_host_params).map_err(|e| {
+                eyre::eyre!("Failed to serialize bridge circuit host params: {}", e)
+            })?;
+            std::fs::write(&path, serialized_params).map_err(|e| {
+                eyre::eyre!("Failed to write bridge circuit host params to file: {}", e)
+            })?;
+            tracing::info!(
+                "Bridge circuit host params with annex written to {:?}",
+                &path
+            );
+        } else if self.use_large_annex {
+            let path = PathBuf::from(
+                "../bridge-circuit-host/bin-files/bch_params_challenge_tx_with_large_annex.bin",
+            );
+            std::fs::create_dir_all(path.parent().unwrap())
+                .map_err(|e| eyre::eyre!("Failed to create directory for output file: {}", e))?;
+            let serialized_params = borsh::to_vec(&bridge_circuit_host_params).map_err(|e| {
+                eyre::eyre!("Failed to serialize bridge circuit host params: {}", e)
+            })?;
+            std::fs::write(&path, serialized_params).map_err(|e| {
+                eyre::eyre!("Failed to write bridge circuit host params to file: {}", e)
+            })?;
+            tracing::info!(
+                "Bridge circuit host params with large annex written to {:?}",
+                &path
+            );
+        } else if self.use_large_output {
+            let path = PathBuf::from(
+                "../bridge-circuit-host/bin-files/bch_params_challenge_tx_with_large_output.bin",
+            );
+            std::fs::create_dir_all(path.parent().unwrap())
+                .map_err(|e| eyre::eyre!("Failed to create directory for output file: {}", e))?;
+            let serialized_params = borsh::to_vec(&bridge_circuit_host_params).map_err(|e| {
+                eyre::eyre!("Failed to serialize bridge circuit host params: {}", e)
+            })?;
+            std::fs::write(&path, serialized_params).map_err(|e| {
+                eyre::eyre!("Failed to write bridge circuit host params to file: {}", e)
+            })?;
+            tracing::info!(
+                "Bridge circuit host params with large output written to {:?}",
+                &path
+            );
+        } else if self.use_large_annex_and_output {
+            let path = PathBuf::from("../bridge-circuit-host/bin-files/bch_params_challenge_tx_with_large_annex_and_output.bin");
+            std::fs::create_dir_all(path.parent().unwrap())
+                .map_err(|e| eyre::eyre!("Failed to create directory for output file: {}", e))?;
+            let serialized_params = borsh::to_vec(&bridge_circuit_host_params).map_err(|e| {
+                eyre::eyre!("Failed to serialize bridge circuit host params: {}", e)
+            })?;
+            std::fs::write(&path, serialized_params).map_err(|e| {
+                eyre::eyre!("Failed to write bridge circuit host params to file: {}", e)
+            })?;
+            tracing::info!(
+                "Bridge circuit host params with large annex and output written to {:?}",
+                &path
+            );
+        }
+        Ok(())
+    }
+
+    pub fn maybe_add_large_test_outputs(
+        &self,
+        mut builder: TxHandlerBuilder,
+    ) -> eyre::Result<TxHandlerBuilder> {
+        // Returns the modified builder
+        // Check if the large annex and output scenario is enabled
+        if self.use_large_annex_and_output {
+            for i in 0..2300 {
+                let mut test_taproot_address: [u8; 32] = [0; 32];
+                let num_to_use: u32 = 30000 + i;
+                let num_to_use_bytes = num_to_use.to_le_bytes();
+                // Last 4 bytes of test_taproot_address will be used to differentiate the outputs
+                test_taproot_address[28..32].copy_from_slice(&num_to_use_bytes);
+                let mut additional_taproot_script_vec = vec![0x51, 0x20];
+                additional_taproot_script_vec.extend_from_slice(&test_taproot_address);
+                let additional_taproot_script =
+                    ScriptBuf::from_bytes(additional_taproot_script_vec);
+                let additional_taproot_txout = TxOut {
+                    value: MIN_TAPROOT_AMOUNT,
+                    script_pubkey: additional_taproot_script,
+                };
+                // Reassign the result of add_output back to builder
+                builder = builder.add_output(UnspentTxOut::from_partial(additional_taproot_txout));
+            }
+            tracing::warn!("Using large annex and output");
+        } else if self.use_large_output {
+            for i in 0..2300 {
+                let mut test_taproot_address: [u8; 32] = [0; 32];
+                let num_to_use: u32 = 30000 + i;
+                let num_to_use_bytes = num_to_use.to_le_bytes();
+                // Last 4 bytes of test_taproot_address will be used to differentiate the outputs
+                test_taproot_address[28..32].copy_from_slice(&num_to_use_bytes);
+                let mut additional_taproot_script_vec = vec![0x51, 0x20];
+                additional_taproot_script_vec.extend_from_slice(&test_taproot_address);
+                let additional_taproot_script =
+                    ScriptBuf::from_bytes(additional_taproot_script_vec);
+                let additional_taproot_txout = TxOut {
+                    value: MIN_TAPROOT_AMOUNT,
+                    script_pubkey: additional_taproot_script,
+                };
+                // Reassign the result of add_output back to builder
+                builder = builder.add_output(UnspentTxOut::from_partial(additional_taproot_txout));
+            }
+            tracing::warn!("Using large output");
+        }
+        Ok(builder)
     }
 }
